+++ grain/_src/python/dataset/dataset_test.py
@@ -1393,5 +1393,20 @@ class GetExecutionSummaryTest(parameterized.TestCase):
     self.assertNotIn(log_value, "".join(logs.output))
 
 
+class GetElementSpecTest(parameterized.TestCase):
+
+  def test_get_element_spec_from_map_dataset(self):
+    ds = dataset.MapDataset.range(10)
+    spec = dataset.get_element_spec(ds)
+    self.assertEqual(spec.shape, ())
+    self.assertEqual(spec.dtype, np.int64)
+
+  def test_get_element_spec_from_iter_dataset(self):
+    ds = dataset.MapDataset.range(10).to_iter_dataset()
+    spec = dataset.get_element_spec(ds)
+    self.assertEqual(spec.shape, ())
+    self.assertEqual(spec.dtype, np.int64)
+
+
 if __name__ == "__main__":
   absltest.main()